requirements-backend.txt
# Moved to backend/requirements.txt
# Keep for backward compatibility
fastapi
uvicorn
python-dotenv
requests
